moved label api
