+++ frontend/lost/src/containers/Labels/LabelTreeTable.jsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react'
 import { CAlert, CBadge, CCard, CCardBody, CTooltip } from '@coreui/react'
 import { ReactFlowProvider } from '@xyflow/react'
 import { FaInfoCircle } from 'react-icons/fa'
-import { useExportLabelTree } from '../../actions/label/label-api'
+import { useExportLabelTree } from '../../api/label'
 import BaseModal from '../../components/BaseModal'
 import { LabelTreeEditor } from './LabelTreeEditor/LabelTreeEditor'
 import { convertLabelTreeToReactFlow } from './LabelTreeEditor/label-tree-util'
@@ -13,7 +13,7 @@ import { createColumnHelper } from '@tanstack/react-table'
 import CoreIconButton from '../../components/CoreIconButton'
 import TableHeader from '../../components/TableHeader'
 import BaseContainer from '../../components/BaseContainer'
-import { useImportLabelTree } from '../../actions/label/label-api'
+import { useImportLabelTree } from '../../api/label'
 import CreateLabelTree from './CreateLabelTree'
 import ErrorBoundary from '../../components/ErrorBoundary'
 import InfoText from '../../components/InfoText'
